pkg: merge `setup_tools` to setup.py (#1527)
* pkg: merge setup_tools

---------
